progress on printf

+++ libpi/include/printf.h
@@ -3,11 +3,11 @@
 
 //#include "rpi.h"
 
-int puts(const char* src);
+size_t puts(const char* src);
 
-int printf(const char* fmt, ...) __attribute__((format (printf, 1, 2)));
-int sprintf(char* buf, const char* fmt, ...) __attribute__((format (printf, 2, 3)));
-int dprintf(void (*write_fun)(uint8_t), const char* fmt, ...) __attribute__((format (printf, 2, 3)));
-int snprintf(char* buf, size_t size, const char* fmt, ...) __attribute__((format (printf, 3, 4)));
+size_t printf(const char* fmt, ...) __attribute__((format (printf, 1, 2)));
+size_t sprintf(char* buf, const char* fmt, ...) __attribute__((format (printf, 2, 3)));
+size_t dprintf(void (*write_fun)(uint8_t), const char* fmt, ...) __attribute__((format (printf, 2, 3)));
+size_t snprintf(char* buf, size_t size, const char* fmt, ...) __attribute__((format (printf, 3, 4)));
 
 #endif
